we got a lot done

+++ pkg/vectorstore/vector_store.go
@@ -2,5 +2,5 @@ package vectorstore
 
 type VectorStorer interface {
 	StoreVector(vector []float32) error
-	QueryVector(vector []float32, k int) ([]string, error)
+	QueryVector(vector []float32, k int64) ([]string, error)
 }
